requirements.txt
boris-spider==1.0.4